paper_formatter/requirements.txt
PyQt5==5.15.9
python-docx==0.8.11
PyPDF2==3.0.1
PyMuPDF==1.23.0
reportlab==4.0.4